+++ app/Http/Controllers/Mahasiswa/SuratKeputusanOrganisasiController.php
@@ -22,14 +22,6 @@ class SuratKeputusanOrganisasiController extends Controller
      */
     public function index()
     {
-        try {
-            return view('mahasiswa.surat_organisasi.index', [
-                'name' => $this->name,
-                'organisasi' => SuratKeputusanOrganisasi::all()
-            ]);
-        } catch(\Illuminate\Database\QueryException $e){
-            return redirect()->back()->with('failed', $e->getMessage());
-        }
 
         try {
             return view('mahasiswa.surat_organisasi.index', [
@@ -39,7 +31,6 @@ class SuratKeputusanOrganisasiController extends Controller
         } catch(\Illuminate\Database\QueryException $e){
             return redirect()->back()->with('failed', $e->getMessage());
         }
-
     }
 
     /**
@@ -56,15 +47,6 @@ class SuratKeputusanOrganisasiController extends Controller
             return redirect()->back()->with('failed', $e->getMessage());
         }
 
-        try {
-            return view('mahasiswa.surat_organisasi.create', [
-                'name' => $this->name,
-                'organisasi' => \App\Models\Organisasi::all()
-            ]);
-        } catch(\Illuminate\Database\QueryException $e){
-            return redirect()->back()->with('failed', $e->getMessage());
-        }
-
     }
 
     /**
@@ -72,7 +54,6 @@ class SuratKeputusanOrganisasiController extends Controller
      */
     public function store(Request $request)
     {
-
         try {
             $validated = Validator::make($request->all(), [
                 'sko_subject' => ['required', 'string'],
@@ -134,15 +115,6 @@ class SuratKeputusanOrganisasiController extends Controller
         } catch(\Illuminate\Database\QueryException $e){
             return redirect()->back()->with('failed', $e->getMessage());
         }
-        try {
-            $surat = $suratKeputusanOrganisasi->find(request()->segment(3));
-            return view('mahasiswa.surat_organisasi.document', [
-                'keputusan' => $surat,
-                'signature' => \App\Models\Signature::leftJoin('jenis_surats','signatures.js_id', '=', 'jenis_surats.js_id')->where('signatures.js_id','=', $surat->js_id)->first()
-            ]);
-        } catch(\Illuminate\Database\QueryException $e){
-            return redirect()->back()->with('failed', $e->getMessage());
-        }
     }
 
     /**
@@ -150,23 +122,18 @@ class SuratKeputusanOrganisasiController extends Controller
      */
     public function edit(SuratKeputusanOrganisasi $suratKeputusanOrganisasi)
     {
-        try {
-            return view('mahasiswa.surat_organisasi.edit', [
-                'name' => $this->name,
-                'keputusan' => $suratKeputusanOrganisasi->find(request()->segment(3)),
-                'organisasi' => \App\Models\Organisasi::all()
-            ]);
-        } catch(\Illuminate\Database\QueryException $e){
-            return redirect()->back()->with('failed', $e->getMessage());
-        }
-        try {
-            return view('mahasiswa.surat_organisasi.edit', [
-                'name' => $this->name,
-                'keputusan' => $suratKeputusanOrganisasi->find(request()->segment(3)),
-                'organisasi' => \App\Models\Organisasi::all()
-            ]);
-        } catch(\Illuminate\Database\QueryException $e){
-            return redirect()->back()->with('failed', $e->getMessage());
+        if ($suratKeputusanOrganisasi->sko_no_surat == null || $suratKeputusanOrganisasi->sko_no_surat_old == null) {
+            try {
+                return view('mahasiswa.surat_organisasi.edit', [
+                    'name' => $this->name,
+                    'keputusan' => $suratKeputusanOrganisasi->find(request()->segment(3)),
+                    'organisasi' => \App\Models\Organisasi::all()
+                ]);
+            } catch(\Illuminate\Database\QueryException $e){
+                return redirect()->back()->with('failed', $e->getMessage());
+            }
+        } else {
+            return redirect()->back()->with('failed', 'You not Have Authority!');
         }
     }
 
@@ -223,12 +190,16 @@ class SuratKeputusanOrganisasiController extends Controller
      */
     public function destroy(SuratKeputusanOrganisasi $suratKeputusanOrganisasi)
     {
-        try {
-            SuratKeputusanOrganisasi::destroy($suratKeputusanOrganisasi->sko_id);
+        if ($suratKeputusanOrganisasi->sko_no_surat == null || $suratKeputusanOrganisasi->sko_no_surat_old == null) {
+            try {
+                SuratKeputusanOrganisasi::destroy($suratKeputusanOrganisasi->sko_id);
 
-            return redirect()->to(route('sko.index'))->with('success', 'Deleted Successfully!');
-        } catch(\Illuminate\Database\QueryException $e){
-            return redirect()->back()->with('failed', $e->getMessage());
+                return redirect()->to(route('sko.index'))->with('success', 'Deleted Successfully!');
+            } catch(\Illuminate\Database\QueryException $e){
+                return redirect()->back()->with('failed', $e->getMessage());
+            }
+        } else {
+            return redirect()->back()->with('failed', 'You not Have Authority!');
         }
     }
 }
